+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGE LOG
 
 ## 1.0.6 - Developing...
+* React.PropTypes replaced by PropTypes
 
 ## 1.0.5
 * changed toast bar design